feat: code control flags
