+++ protocol/common/feature_flags.go
@@ -26,12 +26,4 @@ type FeatureFlags struct {
 
 	// Peersyncing indicates whether we should advertise and sync messages with other peers
 	Peersyncing bool
-
-	// AutoRequestHistoricMessages indicates whether we should automatically request
-	// historic messages on getting online, connecting to store node, etc.
-	AutoRequestHistoricMessages bool
-
-	// CuratedCommunitiesUpdateLoopEnabled indicates whether we should disable the curated communities update loop.
-	// Usually should be disabled in tests.
-	CuratedCommunitiesUpdateLoopEnabled bool
 }
