+++ protocol/messenger_mailserver_cycle.go
@@ -409,7 +409,7 @@ func (m *Messenger) connectToMailserver(ms mailservers.Mailserver) error {
 			signal.SendMailserverAvailable(m.mailserverCycle.activeMailserver.Address, m.mailserverCycle.activeMailserver.ID)
 
 			// Query mailserver
-			if m.config.featureFlags.AutoRequestHistoricMessages {
+			if m.config.codeControlFlags.AutoRequestHistoricMessages {
 				go func() {
 					_, err := m.performMailserverRequest(&ms, func(_ mailservers.Mailserver) (*MessengerResponse, error) {
 						return m.RequestAllHistoricMessages(false, false)
@@ -564,7 +564,7 @@ func (m *Messenger) handleMailserverCycleEvent(connectedPeers []ConnectedPeer) e
 					signal.SendMailserverAvailable(m.mailserverCycle.activeMailserver.Address, m.mailserverCycle.activeMailserver.ID)
 				}
 				// Query mailserver
-				if m.config.featureFlags.AutoRequestHistoricMessages {
+				if m.config.codeControlFlags.AutoRequestHistoricMessages {
 					go func() {
 						_, err := m.RequestAllHistoricMessages(false, true)
 						if err != nil {
